+++ CHANGELOG.md
@@ -1,21 +1,43 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [3.2.1] - 2020-11-05
+
+### Changed 
+- Update to IDS framework v3.2.3.
+- Move self-service and example endpoints to admin API.
+- Improve Dockerfile.
+- Add key- and truststore to example configuration.
+- Add default policy (provide access) to resource on creation. 
+
+### Added
+- Update and delete resources from broker.
+- Add configuration controller for GET and PUT configuration model.
+- Add possibility to set a resource id on create.
+- Add possibility to set a representation id on create.
+- Add a description of how the internal database can be replaced by another.
+- Add .dockerignore file.
+
+### Fixed
+- Add representation.
+- Fix token error in test classes.
+- Fix file path in packaged jar.
+
 ## [3.2.0] - 2020-10-09
 
 ### Changed
-- Change call to BrokerService constructor (parameters changed) in BrokerController
-- Change call to IDSHttpService constructor (parameters changed) in ConnectorRequestServiceImpl
-- Replace ConfigProducer with ConfigurationContainer in IdsUtils, MainController and DescriptionMessageHandler
+- Change call to BrokerService constructor (parameters changed) in BrokerController.
+- Change call to IDSHttpService constructor (parameters changed) in ConnectorRequestServiceImpl.
+- Replace ConfigProducer with ConfigurationContainer in IdsUtils, MainController and DescriptionMessageHandler.
 
 ### Added
-- Autowire ConfigurationContainer in constructor to instantiate Connector, KeyStoreManager or ConfigurationModel (previously directly autowired)
-- Add ClientProvider field to all classes using an OkHttpClient (create instance in constructor from autowired ConfigurationContainer) & replace calls to IDSUtils.getClient() with clientProvider.getClient()
-- Add file URI scheme to paths of KeyStore and TrustStore in config.json
-- Add test classes: SelfDescriptionTest, RequestDescriptionTest, RequestArtifactTest, DescriptionRequestMessageHandlingTest, ArtifactRequestMessageHandlingTest
+- Autowire ConfigurationContainer in constructor to instantiate Connector, KeyStoreManager or ConfigurationModel (previously directly autowired).
+- Add ClientProvider field to all classes using an OkHttpClient (create instance in constructor from autowired ConfigurationContainer) & replace calls to IDSUtils.getClient() with clientProvider.getClient().
+- Add file URI scheme to paths of KeyStore and TrustStore in config.json.
+- Add test classes: SelfDescriptionTest, RequestDescriptionTest, RequestArtifactTest, DescriptionRequestMessageHandlingTest, ArtifactRequestMessageHandlingTest.
 
 ### Removed 
-- IDS Connector certificate file
+- IDS Connector certificate file.
 
 ## [3.1.0] - 2020-09-29